Bridge between LimeSurvey and R

The LimeRick package enables close connection between R and LimeSurvey -- one of the most open source scripts for computer aided web interviewing. Using the LimeSurvey RemoteControl2 JSON-RPC API for two way data exchange allows for:

importing on-line survey responses into R,

adding new responses to the survey directly from R,

automatically accessing surveys and question properties,

implementing passive measurement surveys,

developing data products based on real time declarative data collection.

lsSessionCache

Format

An object of class environment of length 1.

Details

This package is a LimeSurvey RemoteControl 2 JSON-RPC API Client.

The lsSessionCache environment is crated upon the package loading and stores session data like session key or timestamp when the session started. The environment is accessed directly only by package functions not by a package user.

If you need help with your research or commercial projects, feel free to contact me via my homepage contact form: http://www.wais.kamil.rzeszow.pl.

See also