Get survey language properties

Get survey language properties

lsGetLanguageProperties(
  surveyID,
  lang = NULL,
  properties = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)

Arguments

surveyID
ID of the Survey

lang
(optional) Language code for the language to update — if not given the base language of the particular survey is used

properties
array|null (optional) A list with the particular properties to get, default to all attributes

lsAPIurl
(optional) The URL of the LimeSurvey RemoteControl 2 JSON-RPC API

sessionKey
(optional) Authentication token, see lsGetSessionKey()

Value

A list in case of success containing the requested values

References

https://api.limesurvey.org/classes/remotecontrol_handle.html#method_get_language_properties

See also

Examples

if (FALSE) {
  lsGetLanguageProperties("123456")
  lsGetLanguageProperties("123456", lang = "fr")
  lsGetLanguageProperties("123456", properties = list("surveyls_email_confirm"))
  lsGetLanguageProperties("123456", lang = "fr",
                          properties = list("surveyls_email_confirm"))
}