Get properties of a question in a survey

Get properties of a survey. All internal properties of a survey are available.

lsGetQuestionProperties(
  questionID,
  lang,
  properties = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)

Arguments

questionID
ID of the question

lang
Language code for the survey language which affects the question title amongst other properties (Note: The API expects one of the survey languages as part of the request rather than falling back to the default language of the survey. However, you can look up that default language using lsGetSurveyProperties())

properties
(optional) A vector with the particular property names to request, otherwise get all settings

lsAPIurl
(optional) The URL of the LimeSurvey RemoteControl 2 JSON-RPC API

sessionKey
(optional) Authentication token, see lsGetSessionKey()

Value

A list of question properties or a single property

See also

Examples

if (FALSE) {
  lsGetQuestionProperties("13", lang = "en")
  lsGetQuestionProperties("13", lang = "en", properties = list("mandatory"))
}