Return the ids and info of (sub-)questions of a survey/group

Return the ids and info of (sub-)questions of a survey/group

lsListQuestions(
  surveyID,
  lang,
  groupID = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)

Arguments

surveyID
ID of the survey

lang
Language code for the survey language (Note: The API expects one of the survey languages as part of the request rather than falling back to the default language of the survey. However, you can look up that default language using lsGetSurveyProperties())

groupID
(optional) ID of the group to list questions

lsAPIurl
(optional) The URL of the LimeSurvey RemoteControl 2 JSON-RPC API

sessionKey
(optional) Authentication token, see lsGetSessionKey()

Value

The list of questions

See also

Examples

if (FALSE) {
  lsListQuestions("123456", lang = "fr")
  lsListQuestions("123456", lang = "fr", groupID = 1)
}